requirements.txt
PySide6
python-mpv
yt-dlp
requests
